auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
python-multipart
passlib[bcrypt]
bcrypt<4.0.0
psycopg2-binary
email-validator
